FE API: move priceByTransportQuery to the frontend-api package
