+++ project-base/app/src/FrontendApi/Resolver/Price/PriceQuery.php
@@ -6,11 +6,7 @@ namespace App\FrontendApi\Resolver\Price;
 
 use App\FrontendApi\Resolver\Price\Exception\ProductPriceMissingUserError;
 use App\Model\Product\Product;
-use ArrayObject;
-use Shopsys\FrameworkBundle\Model\Pricing\Price;
 use Shopsys\FrameworkBundle\Model\Product\Pricing\ProductPrice;
-use Shopsys\FrameworkBundle\Model\Transport\Transport;
-use Shopsys\FrontendApiBundle\Component\GqlContext\GqlContextHelper;
 use Shopsys\FrontendApiBundle\Model\Resolver\Price\PriceQuery as BasePriceQuery;
 
 /**
@@ -22,51 +18,6 @@ use Shopsys\FrontendApiBundle\Model\Resolver\Price\PriceQuery as BasePriceQuery;
  */
 class PriceQuery extends BasePriceQuery
 {
-    /**
-     * @param \App\Model\Transport\Transport $transport
-     * @param string|null $cartUuid
-     * @param ?\ArrayObject $context
-     * @return \Shopsys\FrameworkBundle\Model\Pricing\Price
-     */
-    public function priceByTransportQuery(
-        Transport $transport,
-        ?string $cartUuid = null,
-        ?ArrayObject $context = null,
-    ): Price {
-        $cartUuid = $cartUuid ?? GqlContextHelper::getCartUuid($context);
-
-        /** @var \App\Model\Customer\User\CustomerUser|null $customerUser */
-        $customerUser = $this->currentCustomerUser->findCurrentCustomerUser();
-
-        if ($customerUser === null && $cartUuid === null) {
-            return parent::priceByTransportQuery($transport);
-        }
-
-        $cart = $this->cartApiFacade->findCart($customerUser, $cartUuid);
-
-        if ($cart === null) {
-            return parent::priceByTransportQuery($transport);
-        }
-
-        $domainId = $this->domain->getId();
-        $currency = $this->currencyFacade->getDomainDefaultCurrencyByDomainId($domainId);
-        $orderPreview = $this->orderPreviewFactory->create(
-            $currency,
-            $domainId,
-            $cart->getQuantifiedProducts(),
-            $transport,
-            null,
-            $customerUser,
-        );
-
-        return $this->transportPriceCalculation->calculatePrice(
-            $transport,
-            $currency,
-            $orderPreview->getProductsPrice(),
-            $domainId,
-        );
-    }
-
     /**
      * {@inheritdoc}
      */
